requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.6.2.post1
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
bleach==6.2.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clldutils==3.24.0
cloudpathlib==0.20.0
cn2an==0.5.22
colorama==0.4.6
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.0
coqpit==0.0.17
coqui-tts==0.24.3
coqui-tts-trainer==0.1.6
csvw==3.5.1
cutlet==0.4.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dlinfo==1.2.1
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.4
fastjsonschema==2.20.0
ffmpeg==1.4
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.2.0
fugashi==1.1.2
gradio==5.5.0
gradio_client==1.4.2
grpcio==1.67.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.14.0
hjson==3.1.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
inflect==7.4.0
ipython==8.12.3
isodate==0.7.2
jaconv==0.4.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.3
jiter==0.7.0
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.7
langcodes==3.4.1
langid==1.1.6
language-tags==1.2.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mecab-python3==1.0.10
mistune==3.0.2
mojimoji==0.0.13
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.2.1
ninja==1.11.1.1
num2words==0.5.13
numba==0.60.0
number2text==0.0.1
numpy==1.26.4
omegaconf==2.3.0
openai==1.54.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
phonemizer==3.3.0
pickleshare==0.7.5
pillow==10.2.0
pipreqs==0.5.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
proces==0.1.7
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pylatexenc==2.10
pynvml==11.5.3
PyOpenGL==3.1.7
pyopenjtalk==0.3.4
pyparsing==3.2.0
pypinyin==0.53.0
pysbd==0.3.4
PySide2==5.15.2.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.21.0
ruff==0.7.3
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
segments==2.2.1
semantic-version==2.10.0
shellingham==1.5.4
shiboken2==5.15.2.1
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
SudachiDict-core==20241021
SudachiPy==0.6.8
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.12.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typeguard==4.4.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.2
win32-setctime==1.1.0
wrapt==1.16.0
yarg==0.1.9
yarl==1.17.1
